chore: move team store
